+++ src/index.test.ts
@@ -1,109 +1,13 @@
-import { promises as fs } from 'fs'
-import { tmpdir } from 'os'
-import { dirname, resolve } from 'path'
+import {
+    TEST_PACKAGE_NAME,
+    cleanUp,
+    createFile,
+    declareDependencies,
+    prepareTempDirectory,
+    prepareTempMonorepoDirectory,
+} from './testUtils'
 
 import validateDependencies from '.'
-
-const TEST_PACKAGE_NAME = 'testpackage'
-
-async function prepareTempDirectory(
-    cwd?: string,
-    packageName?: string,
-): Promise<string> {
-    let tempRoot
-    if (cwd) tempRoot = cwd
-    else tempRoot = await fs.mkdtemp(`${tmpdir()}/`)
-
-    // A lockfile is needed, but its content don't matter.
-    await createFile(tempRoot, 'yarn.lock', '')
-
-    const packageJsonTemplate = {
-        name: packageName ?? TEST_PACKAGE_NAME,
-        dependencies: {},
-        devDependencies: {},
-        peerDependencies: {},
-    }
-    // Basic package.json without dependencies
-    await createFile(
-        tempRoot,
-        'package.json',
-        JSON.stringify(packageJsonTemplate),
-    )
-
-    return tempRoot
-}
-
-async function prepareTempMonorepoDirectory(): Promise<Set<string>> {
-    const tempRoot = await fs.mkdtemp(`${tmpdir()}/`)
-
-    // A lockfile is needed, but its content don't matter.
-    await createFile(tempRoot, 'yarn.lock', '')
-
-    const packageJsonTemplate = {
-        name: TEST_PACKAGE_NAME,
-        private: true,
-        workspaces: ['packages/*'],
-        dependencies: {},
-        devDependencies: {},
-        peerDependencies: {},
-    }
-    // Basic package.json without dependencies
-    await createFile(
-        tempRoot,
-        'package.json',
-        JSON.stringify(packageJsonTemplate),
-    )
-
-    const createdPaths = [tempRoot]
-    const subPackages = ['mono1', 'mono2', 'mono3']
-    await Promise.all(
-        subPackages.map(async (packageName) => {
-            const packagePath = await prepareTempDirectory(
-                `${tempRoot}/packages/${packageName}`,
-                packageName,
-            )
-            createdPaths.push(packagePath)
-        }),
-    )
-    return new Set(createdPaths)
-}
-async function declareDependencies(
-    cwd: string,
-    dependencies: Map<string, string>,
-): Promise<string> {
-    const currentPackageJson = JSON.parse(
-        await fs.readFile(resolve(cwd, 'package.json'), { encoding: 'utf8' }),
-    )
-
-    for (const dependency of dependencies.entries()) {
-        const [packageName, dependencyClass] = dependency
-
-        if (dependencyClass === 'prod')
-            currentPackageJson.dependencies[packageName] = '*'
-        else if (dependencyClass === 'dev')
-            currentPackageJson.devDependencies[packageName] = '*'
-        else if (dependencyClass === 'peer')
-            currentPackageJson.peerDependencies[packageName] = '*'
-    }
-
-    await createFile(cwd, 'package.json', JSON.stringify(currentPackageJson))
-
-    return currentPackageJson
-}
-
-async function createFile(
-    cwd: string,
-    path: string,
-    data: string,
-): Promise<string> {
-    const filePath = resolve(cwd, path)
-    const parentDirectory = dirname(filePath)
-
-    await fs.mkdir(parentDirectory, { recursive: true })
-    await fs.writeFile(filePath, data, { encoding: 'utf8' })
-
-    return filePath
-}
 
 describe('GhostImports', () => {
     describe('High-level use cases (single project)', () => {
@@ -114,7 +18,7 @@ describe('GhostImports', () => {
         })
 
         afterEach(async () => {
-            await fs.rmdir(workspacePath, { recursive: true })
+            await cleanUp(workspacePath)
         })
 
         it('detects require imports correctly', async () => {
@@ -225,7 +129,7 @@ describe('GhostImports', () => {
         })
 
         afterEach(async () => {
-            await fs.rmdir(workspacePath, { recursive: true })
+            await cleanUp(workspacePath)
         })
 
         it('produces report with correct package analysis', async () => {
